refactor: extract /config, /export, /import into internal/server/config
Phase 4a of the webhook split (issue #250). Moves the config snapshot
and YAML import/export HTTP surface plus the methods exposed for the
MCP get_config / export_config / import_config tools into a new
dedicated package.

Architecture:
  - internal/server/config/Handler holds /config (read-only snapshot
    with secrets redacted), /export (YAML dump of CRUD-mutable
    sections), and /import (YAML upsert with merge/replace modes), plus
    ConfigJSON / ExportYAML / ImportYAML methods for the MCP layer.
  - The handler boots in cfg-only mode (NewServer) so /config works for
    tests and pre-WithStore daemons; SetDB wires the database for
    /export and /import. CRUD writes flow through the same shared
    server.WriteCoordinator so the lock + reload epoch stay coherent
    with the other domain handlers.
  - webhook.Server constructs the handler in NewServer, exposes it via
    Cfg(), and cmd/agents wires it as ConfigBytes / ConfigImport for
    MCP tools.
  - All apiConfigJSON / exportYAML wire types move with the handler.

webhook drops ~415 lines: api.go is gone entirely, crud.go shrinks to
just the storeWriteErr / storeErrStatus / reloadCron helpers shared
across the still-in-webhook routes (/status, /run, /webhooks/github).
The api_test.go file now hits routes through buildHandler() instead of
calling unexported handlers directly.

What's left in webhook: HTTP server lifecycle, /webhooks/github (HMAC
+ delivery dedupe + event parsers), /status, /run, the Anthropic
proxy mount, UI static mount, and MCP mount. Phase 4b will move the
HTTP server itself out of webhook so the package becomes pure
GitHub-webhook handling.
