+++ cmd/agents/main.go
@@ -168,8 +168,8 @@ func run() error {
 		Observe:       obs,
 		DispatchStats: engine,
 		Memory:        &sqliteMcpReader{db: db},
-		ConfigBytes:   srv,
-		ConfigImport:  srv,
+		ConfigBytes:   srv.Cfg(),
+		ConfigImport:  srv.Cfg(),
 		AgentWrite:    srv.Fleet(),
 		SkillWrite:    srv.Fleet(),
 		BackendWrite:  srv.Fleet(),
